fix(fromReferenceSeconds): throw if the specified count of seconds is not a finite value
-

BREAKING CHANGE: `fromReferenceSeconds` now throws if the specified count of seconds is not a finite
value. Previously it would return an invalid Time object with the same non-finite value as its
hours, minutes and seconds.

BREAKING CHANGE: `time` now throws if any of the component parts of the specified time are not
finite values. Previously it would return an invalid Time object with non-finite values as its
hours, minutes and/or seconds.

BREAKING CHANGE: `normalize` now throws if any of the component parts of the specified time are
not finite values. Previously it would return an invalid Time object with non-finite values as its
hours, minutes and/or seconds.